requirements.txt
requests==2.31.0
numpy>=2.2.4
sentence-transformers>=3.0.0
matplotlib==3.8.4
python-dotenv==1.0.1
endee